관리자 회원삭제(일반, 크루장위임, 크루원) 변경수정

+++ HipBus/src/handler/admin/AdminMemberEjectPro.java
@@ -1,6 +1,8 @@
 package handler.admin;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
@@ -27,14 +29,35 @@ public class AdminMemberEjectPro implements CommandHandler {
 	public ModelAndView process(HttpServletRequest request, HttpServletResponse response) throws HandlerException {
 		String email = request.getParameter("email");
 		
-		List<String> list = adminDao.crewLeader(email);		
-		for(int i=0;i<list.size();i++){
-			List<MemberDto> clist =adminDao.crewId( list.get(i));
-			adminDao.updateLeader(clist.get(0).getEmail());
+		List<String> list = adminDao.crewLeader(email);	
+		
+		int result = 0;
+		int resultcheck = 0;
+		
+		if(list.size() == 0){
+			result = adminDao.deleteMember(email);
+			resultcheck = adminDao.deleteDriver(email);
 		}
-				
-		int result = adminDao.deleteMember(email);
-		int resultcheck = adminDao.deleteDriver(email);
+
+		if(list.size() != 0){
+			for(int i=0;i<list.size();i++){
+				List<MemberDto> clist = null;
+				clist = adminDao.crewId( list.get(i) );
+				Map<String,String> map = new HashMap<String,String>();
+				if(clist.isEmpty()){
+					result = adminDao.deleteMember(email);
+					resultcheck = adminDao.deleteDriver(email);
+				} else {
+					map.put("email", clist.get(0).getEmail());
+					map.put("crewid", list.get(i));
+					
+					adminDao.updateLeader(map);
+					result = adminDao.deleteMember(email);
+					resultcheck = adminDao.deleteDriver(email);
+				}
+			}
+		}
+
 		request.setAttribute("result", result);
 		request.setAttribute("resultcheck", resultcheck);
